requirements.txt
python-telegram-bot[all]==20.7
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pillow==10.1.0